backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
distlib==0.4.0
dotenv==0.9.9
filelock==3.25.2
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
platformdirs==4.9.4
psycopg2==2.9.11
PyJWT==2.12.1
python-discovery==1.2.1
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
uv==0.9.18
virtualenv==21.2.0
Werkzeug==3.1.7
